new stuff

+++ Test/MyMessageHandler.cpp
@@ -121,7 +121,8 @@ MyMessageHandler::print()
         int numExtremePointsSaved = feasibleExtremePoints_.size();
         if ( numExtremePointsSaved>=10 ) {
           feasibleExtremePoints_.pop_back();
-          assert( feasibleExtremePoints_.size() == numExtremePointsSaved-1 );
+          assert( feasibleExtremePoints_.size() == 
+		  (unsigned int) numExtremePointsSaved-1 );
         };
 
       }
